Add functions for array busy-flags handling
The column elements cannot be handled like the elements in the columns -
they can not be replaced by the last entry in the array. Columns are
therefore expected to be handled by marking removed spots as not busy
since the column arrays or pointer to pointer arrays are never iterated
upon with a for-loop. Functions to deal with this are therefore added.